next

+++ src/Expressions.java
@@ -4,7 +4,7 @@ public class Expressions {
 	public static void main(String[] args) {
 		int friends = 10;
 		
-		friends = (double)friends / 3;
+		friends = friends / 3;
 		
 		System.out.println(friends);
 	}
